update support fireredasr_aed_l
